Move decorators from auth to shared folder
